requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
groq==0.11.0
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2

# AI Service Dependencies
openai==1.3.0
google-generativeai==0.3.0